pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "gitpython>=3.1.46",
    "markdownify>=1.2.2",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
]
